Rename micro-service name to manager
