+++ .github/workflows/build-and-push.yml
@@ -32,7 +32,7 @@ jobs:
         if: github.event_name == 'pull_request' || github.event_name == 'push'
         uses: docker/build-push-action@6.18.0
         with:
-          context: .
+          context: manager
           push: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' && github.repository == 'platform-engineering-org/afula' }}
-          tags: ghcr.io/${{ github.repository_owner }}/afula:latest
+          tags: ghcr.io/${{ github.repository_owner }}/afula-manager:latest
           platforms: linux/amd64,linux/arm64/v8
