+++ .github/workflows/build-and-scan.yml
@@ -21,16 +21,16 @@ jobs:
         if: github.event_name == 'pull_request_target'
         uses: docker/build-push-action@6.18.0
         with:
-          context: .
+          context: manager
           load: true
           push: false
-          tags: afula:pr-${{ github.event.pull_request.number }}
+          tags: afula-manager:pr-${{ github.event.pull_request.number }}
           platforms: linux/amd64
 
       - name: Scan image for vulnerabilities
         uses: aquasecurity/trivy-action@0.31.0
         with:
-          image-ref: afula:pr-${{ github.event.pull_request.number }}
+          image-ref: afula-manager:pr-${{ github.event.pull_request.number }}
           format: table
           output: trivy-results.txt
           exit-code: 0
@@ -50,7 +50,7 @@ jobs:
         with:
           issue-number: ${{ github.event.pull_request.number }}
           comment-author: 'github-actions[bot]'
-          body-includes: '### 🔒 Trivy Scan Results for `afula:pr-${{ github.event.pull_request.number }}`'
+          body-includes: '### 🔒 Trivy Scan Results for `afula-manager:pr-${{ github.event.pull_request.number }}`'
 
       - name: Post Trivy scan result as PR comment
         uses: peter-evans/create-or-update-comment@v4.0.0
@@ -59,7 +59,7 @@ jobs:
           comment-id: ${{ steps.get-comment.outputs.comment-id }}
           edit-mode: replace
           body: |
-            ### 🔒 Trivy Scan Results for `afula:pr-${{ github.event.pull_request.number }}`
+            ### 🔒 Trivy Scan Results for `afula-manager:pr-${{ github.event.pull_request.number }}`
             ```
             ${{ steps.read-scan.outputs.scan_output }}
             ```
