+++ .github/workflows/ci.yml
@@ -22,7 +22,7 @@ jobs:
           python-version: '3.13'
 
       - name: Install dependencies
-        run: pip install -r dev-requirements.txt -r requirements.txt
+        run: pip install -r dev-requirements.txt -r manager/requirements.txt
 
       - name: Run pre-commit
         run: pre-commit run --all-files
