+++ .github/workflows/deploy-kind.yml
@@ -20,10 +20,10 @@ jobs:
       - name: Build
         uses: docker/build-push-action@6.18.0
         with:
-          context: .
+          context: manager
           load: true
           push: false
-          tags: ghcr.io/${{ github.repository_owner }}/afula:latest
+          tags: ghcr.io/${{ github.repository_owner }}/afula-manager:latest
           platforms: linux/amd64
 
       - name: Create Kind cluster
@@ -35,7 +35,7 @@ jobs:
 
       - name: Apply manifests
         run: |
-          kind load docker-image ghcr.io/${{ github.repository_owner }}/afula:latest --name afula
-          kubectl apply -f deploy/postgres-secret.yaml
-          kubectl apply -f deploy/postgres.yaml
-          kubectl apply -f deploy/afula.yaml
+          kind load docker-image ghcr.io/${{ github.repository_owner }}/afula-manager:latest --name afula
+          kubectl apply -f manager/deploy/postgres-secret.yaml
+          kubectl apply -f manager/deploy/postgres.yaml
+          kubectl apply -f manager/deploy/manager.yaml
