+++ manager/Dockerfile
@@ -9,7 +9,7 @@ COPY requirements.txt .
 
 RUN pip install -r requirements.txt
 
-COPY ./afula .
+COPY . .
 
 EXPOSE 5000
