+++ manager/Makefile
