+++ manager/__init__.py
