+++ manager/deploy/manager.yaml
@@ -1,22 +1,22 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: afula
+  name: manager
   labels:
-    app: afula
+    app: manager
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app: afula
+      app: manager
   template:
     metadata:
       labels:
-        app: afula
+        app: manager
     spec:
       containers:
-      - name: afula
-        image: ghcr.io/platform-engineering-org/afula:latest
+      - name: manager
+        image: ghcr.io/platform-engineering-org/afula-manager:latest
         imagePullPolicy: Never
         ports:
         - containerPort: 5000
@@ -27,10 +27,10 @@ spec:
 apiVersion: v1
 kind: Service
 metadata:
-  name: afula
+  name: manager
 spec:
   selector:
-    app: afula
+    app: manager
   ports:
   - port: 5000
     targetPort: 5000
