+++ manager/deploy/postgres-secret.yaml
