+++ manager/requirements.txt
