+++ manager/templates/register_form.html
