+++ manager/templates/repos.html
