+++ manager/templates/repositories.html
